The `SelectComponent` doesn't block page rendering with big collection.
The dropdown observes any changes to the `optionsLabelPath`. This can be
pretty slow with a big collection (several seconds) and it would get
computed even if the dropdown would never be opened.

This commit allows to do the computation only when the dropdown is
actually opened.

This is not a perfect solution as the slowness will still appear the
first time the dropdown is opened ( or if the content changes ).

Code by @bigsley and @cyril-sf